requirements.txt
pandas
matplotlib
seaborn
scikit-learn
tensorflow
keras
opencv-python
gradio
Pillow